Image detection and loading
If an image is detected, neither the video or audio thread is created. Instead, the image is simply loaded
into MediaPlayer::frame and the MediaPlayer::render_loop is run

Additionally, basic audio stream support is present. If an audio file is detected, the video thread
is not created, but the audio thread runs. In the future, this will also include a rendering of the audio
waveform to display to the user. However, this branch is not necessarily targeted toward creating audio functionality.
I only added the audio functionality because it seemed to correlate directly with the functionality needed to create image support.