Clear audio buffer on main thread as well as audio thread
We have to clear the audio buffer on the main thread since this
effects how desyncing is calculated by the MediaFetcher. This
prevents the main thread from spamming jump_to_time on every
iteration whenever the
MediaFetcher becomes out of sync. We then
clear on the audio thread like before
so that the frames inside of the audio
buffer are ensured to remain valid to the current media time.